Make Mu a normal compile-time dependency (#20)
I just tried publishing the plugin locally and using it in a project,
and the `muSrcGen` task threw a ClassNotFoundException.

Mu will be on the classpath of an end-user's Mu application, but it
won't be on the classpath of the sbt plugin. So I don't think a
"provided" dependency makes sense here.

Confirmed locally that making it a normal dependency fixes the issue.