Switch from skeuomorph to ScalaPB for protobuf srcgen
This is the first step towards Scala 3 support for Mu.

* Add a dependency on the ScalaPB sbt-protoc plugin
* Implement a custom protoc plugin to generate the Mu service trait
    * Use ScalaPB's standard functionality to generate the message
      classes, and our custom plugin for the Mu service
* Extend the scripted test to verify the generated code
* Remove the skeuomorph-based protobuf source generator
* Split the sbt plugin's `muSrcGen` task into more granular tasks so we
  can re-use parts of it for both Avro and protobuf srcgen
* Conditionally enable either our source generator (used only for Avro
  now) or ScalaPB, depending on the value of `muSrcGenIdlType`
* Add sbt-buildinfo because it's required for the protoc plugin

The next step will be to extend the protoc plugin to generate the
service trait's companion object, containing all the helper methods that
are currently macro-generated.

This branch will not build in CI yet because it depends on https://github.com/higherkindness/mu-scala/pull/1442