Remove protobuf custom marshaller imports
These are not needed anymore, because we have switched from pbdirect to
ScalaPB for protobuf serialization. These imports were to bring pbdirect
type class instances into scope.

Because ScalaPB takes care of serialization for us, there should be no
need to customise protobuf serialization at all (and there is not really
any mechanism for doing so).