Remove support for generating un-tagged BigDecimals
Mu's custom codec for raw BigDecimals has been deprecated for over 3
years, for good reason.

If Avrohugger generates a BigDecimal, it means the original Avro type
was a decimal, with a maximum precision and scale (e.g. `decimal(6, 2)`). If we
throw away that precision and scale information and use Mu's raw
BigDecimal codec, we will produce wire values that don't match the Avro
schema. This is pretty dangerous.

I considered 3 options:

1. Keep the deprecated codec and port it to Scala 3. I don't want to do
   this, for the reason outlined above.
2. Remove the deprecated codec. Serialization of BigDecimals would fall
   back to Avro4s's built-in codec, which blindly assumes certain values
   for scale and precision. Of course, those values might not match the
   values provided in the Avro schema, so this is just as dangerous as
   option 1.
3. Remove support for raw BigDecimals entirely. Configure Avrohugger to
   always tag the BigDecimal with its maximum precision and scale,
   allowing us to reconstruct the `decimal` logical type and serialize
   appropriately.